+++ src/url-loader/fs-url-loader.ts
@@ -37,7 +37,7 @@ export class FSUrlLoader implements UrlLoader {
     return this._isValid(urlObject, pathname);
   }
 
-  _isValid(urlObject: Url, pathname: string) {
+  private _isValid(urlObject: Url, pathname: string) {
     return (urlObject.protocol === 'file' || !urlObject.hostname) &&
         !pathname.startsWith('../');
   }
@@ -64,4 +64,35 @@ export class FSUrlLoader implements UrlLoader {
     }
     return this.root ? pathlib.join(this.root, pathname) : pathname;
   }
-};
+
+  async readDirectory(pathFromRoot: string, deep?: boolean): Promise<string[]> {
+    const files = await new Promise<string[]>((resolve, reject) => {
+      fs.readdir(
+          pathlib.join(this.root, pathFromRoot),
+          (err, files) => err ? reject(err) : resolve(files));
+    });
+    const results = [];
+    const subDirResultPromises = [];
+    for (const basename of files) {
+      const file = pathlib.join(pathFromRoot, basename);
+      const stat = await new Promise<fs.Stats>(
+          (resolve, reject) => fs.stat(
+              pathlib.join(this.root, file),
+              (err, stat) => err ? reject(err) : resolve(stat)));
+      if (stat.isDirectory()) {
+        if (deep) {
+          subDirResultPromises.push(this.readDirectory(file, deep));
+        }
+      } else {
+        results.push(file);
+      }
+    }
+    const arraysOfFiles = await Promise.all(subDirResultPromises);
+    for (const dirResults of arraysOfFiles) {
+      for (const file of dirResults) {
+        results.push(file);
+      }
+    }
+    return results;
+  }
+}