+++ src/url-loader/url-loader.ts
@@ -14,4 +14,13 @@ export interface UrlLoader {
    * This should only be called if `canLoad` returns `true` for `url`.
    */
   load(url: string): Promise<string>;
+
+  /**
+   * Lists files in a directory in the current project.
+   *
+   * @param path A relative path to a directory to read.
+   * @param deep If true, lists files recursively. Returned paths are
+   *     relative to `url`.
+   */
+  readDirectory?(path: string, deep?: boolean): Promise<string[]>;
 }